Optimize memory for tall-matrix case for FullPivLu
Previously, the L matrix would always be a mxm matrix, and the U matrix
would be a mxn matrix. This is because the unpack method used the
self.lu matrix exclusively to store the U matrix. In order to avoid
wasting large amounts of memory, the internal storage is repurposed
to store either the L or U matrix, whichever can fit into the mxn size.
The other matrix is allocated as a square matrix.

Also removed the Mul trait for FullPivLu, and moved that trait
to a private internal function.